refactor(mentor): 7주차 코드에서 디렉토리 구조 변경
