+++ src/assets/styles/folderPage/AddLink.css
@@ -15,7 +15,7 @@
   height: 4.31rem;
   margin: 3.75rem 0 5.62rem 0;
   background: var(--linkbrary-white, #fff);
-  background-image: url("../../assets/link.svg");
+  background-image: url("../../icons/header/link.svg");
   background-repeat: no-repeat;
   background-size: 1.25rem;
   background-position: 1.25rem;
