+++ src/assets/styles/folderPage/NoLinks.css
