+++ src/assets/styles/search/SearchBar.css
@@ -2,7 +2,7 @@
   width: 100%;
   height: 3.375rem;
   background: #f5f5f5;
-  background-image: url("../../assets/Search.svg");
+  background-image: url("../../icons/search/Search.svg");
   background-repeat: no-repeat;
   background-size: 1rem;
   background-position: 1rem;
