+++ src/components/common/CardInfo.js
