+++ src/components/common/Logo.js
@@ -1,4 +1,4 @@
-import logo from "../../assets/logo.png";
+import logo from "../../assets/icons/header/logo.png";
 export function Logo() {
   return (
     <>
