+++ src/components/common/SearchBar.js
@@ -1,4 +1,4 @@
-import "./SearchBar.css";
+import "../../assets/styles/search/SearchBar.css";
 export function SearchBar() {
   return (
     <>
