+++ src/components/common/Sns.js
@@ -3,7 +3,7 @@ export function Sns({ footerSnsData }) {
 
   return (
     <>
-      <a href={url} target="_blank">
+      <a href={url} target="_blank" rel="noreferrer">
         <img src={img} alt={name} />
       </a>
     </>
