+++ src/components/common/folderPage/AddLink.js
@@ -1,4 +1,4 @@
-import "./AddLink.css";
+import "../../../assets/styles/folderPage/AddLink.css";
 function AddLink() {
   return (
     <div className="add-link-container">
