+++ src/components/common/folderPage/NoLinks.js
@@ -1,4 +1,4 @@
-import "./NoLinks.css";
+import "../../../assets/styles/folderPage/NoLinks.css";
 export function Nolinks({ msg }) {
   return <div className="no-links">{msg}</div>;
 }
