+++ src/pages/SharedPage.js
@@ -1,11 +1,11 @@
-import Header from "../header/Header.js";
-import Footer from "../footer/Footer.js";
+import Header from "../components/common/Header.js";
+import Footer from "../components/common/Footer.js";
 import { useEffect, useState } from "react";
-import { getData } from "../api/api";
-import MainHeader from "../main/MainHeader.js";
-import SearchBar from "../main/SearchBar.js";
-import CardWrapper from "../main/CardWrapper.js";
-import "../main/Main.css";
+import { getData } from "../utils/api.js";
+import MainHeader from "../components/common/sharedPage/MainHeader.js";
+import SearchBar from "../components/common/SearchBar.js";
+import CardWrapper from "../components/common/CardWrapper.js";
+import "../assets/styles/header/MainHeader.css";
 
 function SharedPage() {
   const [profileDatas, setProfileDatas] = useState({
